npm audit is very upset about the state of the repo, fix it. There doesn't appear to be much reason to avoid this, and it solves 85 active known vulnerabilities in the various packages pulled into rambox.
WARNING: This includes electron 2.0.2. This is probably a good idea, it
works in my testing, there don't appear to be breaking changes, but this
needs extra testing, I'm sure.

Extra WARNING: This also includes a sencha cmd app update for good measure.
An ExtJS framework update is left for another time.