Inbox: only update unread count if the current view is the Inbox view.
Currently the unread badges are updated in any Inbox view (such as Done,
Snoozed or any folder). These are not unread/new messages and as such
should not count towards the badge.

This patch works around that issue by checking if the current view is
the Inbox view, and not updating the unread count otherwise.