Inbox: only update notification badge for Inbox view.
Currently the unread badges are updated in any view (such as Done,
Snoozed or any folder). These are not unread/new messages and as such
should not count towards the badge.

This patch works around that issue by checking if the current view is
the Inbox view, and not updating the unread count otherwise.

See: https://github.com/ramboxapp/community-edition/issues/1136,
https://github.com/ramboxapp/community-edition/pull/1853